Move registercommand back to Discord main class
